Update README.md
Add a note on cabal-1.22.5, which seems to be prone to errors when installing packages. I used 1.22.5, since it is in the matrix of the Haskell Platform for the latest GHC 7.10 release, hence it was my preferred choice.
It didn't work at all on my fresh setup on a Raspberry Pi and updating to cabal-1.22.9 solved the problem for me. Answers on github bugs and Stack Overflow suggest that the relevant fix is already included in cabal 1.22.6.